ARM64: enforce explicit host/token width boundaries

+++ port/shim/PR/os.h
@@ -34,7 +34,7 @@
 #    define OS_K0_TO_PHYSICAL(x) ((u32)((uintptr_t)(x) - (uintptr_t)0x70000000u))
 
 #    undef OS_PHYSICAL_TO_K0
-#    define OS_PHYSICAL_TO_K0(x) ((void *)(uintptr_t)(u32)(x))
+#    define OS_PHYSICAL_TO_K0(x) ((void *)(uintptr_t)(u32)(uintptr_t)(x))
 
 #else
 #    include <PR/os.h>
