+++ src/game/title.c
@@ -555,7 +555,7 @@ void initializeGunBarrelIntro(u8 *gfxBuffer, s32 bufferSize)
      * N64 K0 value from the physical offset; seg_addr() passes it through to
      * the KSEG0 mirror at 0x80000000 (segments 7/8 are never registered, so
      * the unmarked-segment path is skipped). */
-    sub_GAME_7F01BFF8(gunbarrelgfxListPointer, (Vtx *)(OS_K0_TO_PHYSICAL((void *)barrelDisplayListPtr) | 0x80000000u), 0x1E);
+    sub_GAME_7F01BFF8(gunbarrelgfxListPointer, (Vtx *)(uintptr_t)(OS_K0_TO_PHYSICAL((void *)barrelDisplayListPtr) | 0x80000000u), 0x1E);
 #else
     sub_GAME_7F01BFF8(gunbarrelgfxListPointer, barrelDisplayListPtr + 0x80000000, 0x1E);
 #endif
